+++ app/containers/ChartMobile.js
@@ -1,0 +1,12 @@
+import { connect } from 'react-redux';
+import ChartMobileComponent from '../components/ChartMobile';
+
+const mapStateToProps = (state) => {
+    return {
+        sensorData: state.sensorDetail.datapoints
+    }
+};
+
+const ChartMobile = connect(mapStateToProps)(ChartMobileComponent);
+
+export default ChartMobile