+++ app/pages/Search.jsx
@@ -21,7 +21,7 @@ class Search extends Component {
                 <Content>
                     <div className={styles.bodymap}>
                         <Grid className={styles.noPadding}>
-                            <Cell col={2}>
+                            <Cell col={3}>
                                 <List className={styles.list}>
                                     <FilterSelection/>
                                 </List>
@@ -29,7 +29,7 @@ class Search extends Component {
                                     <DownloadButtons/>
                                 </div>
                             </Cell>
-                            <Cell col={10}>
+                            <Cell col={9}>
                                 <div className={styles.rightMap} >
                                     <SearchMap />
                                 </div>