+++ app/reducers/sensorDetail.js
@@ -1,4 +1,6 @@
-import { SELECT_SENSOR, RECEIVE_SENSOR, UPDATE_DETAIL } from '../actions'
+import {
+    SELECT_SENSOR, RECEIVE_SENSOR, UPDATE_DETAIL
+} from '../actions';
 
 const defaultState = {id: null, datapoints: []};
 
@@ -27,7 +29,7 @@ const sensorDetail = (state:backendsState = defaultState, action) => {
 
 //TODO: sum to average
 function groupBy(array, col, value) {
-    var r = [], o = {};
+    let r = [], o = {};
     array.forEach(function (a) {
         if (!o[a[col]]) {
             o[a[col]] = {};
@@ -38,11 +40,17 @@ function groupBy(array, col, value) {
         o[a[col]][value] += +a[value];
     });
     return r;
-};
+}
 
 function collectdata(data){
+
+    // Handle no return from API
+    if (data.length <=0) {
+        return [];
+    }
+
     //console.log(data[0]);
-    for(var key in data[0].properties){
+    for(let key in data[0].properties){
         data[0].properties[key] = groupBy(data[0].properties[key], 'label', 'average');
     }
     return data[0].properties;