+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GLGVO-376](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-376)
 - Filtering available dates that show up on the search page based on previous filters.
   [GEOD-1024](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1024)
+- Added Mobile view options
+  [IMLCZO-206](https://opensource.ncsa.illinois.edu/jira/browse/IMLCZO-206)
+
+### Fixed
 
 ### Changed